+++ lecture1(14.03)/CommitOrder2/CommitOrder2/Program.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-class Program
+﻿class Program
 {
 
     static void Main()
@@ -10,35 +8,33 @@ class Program
         string? name = Console.ReadLine();
 
         Console.WriteLine("Введите количество вашего товара:");
-        
+        int? productQuantity = 0;
         try
         {
-            int? productQuantity = Convert.ToInt32(Console.ReadLine());
-
-            Console.WriteLine("Введите название вашего товара:");
-            string? product = Console.ReadLine();
-
-            Console.WriteLine("Введите адрес доставки:");
-             string? address = Console.ReadLine();
-            Console.WriteLine("\nПодтверждение введенных данных:");
-            Console.WriteLine($"Здравствуйте, {name}, вы заказали {productQuantity} {product} на адрес {address}, верно?");
-            Console.WriteLine("Это верно? -- да, нет");
-            string answer = Console.ReadLine();
-            switch (answer)
-            {
-                case "да":
-                    Console.WriteLine($"{name}, ваш заказ: {product} в колличестве: {productQuantity} оформлен, ожидайте доставку по адресу: {address} к {todayDate}");
-                    break;
-                default:
-                    Console.WriteLine("Пожалуйста, перепроверьте свои данные");
-                    break;
-            }
+           productQuantity = Convert.ToInt32(Console.ReadLine());
         }
         catch (FormatException)
         {
             Console.WriteLine("Некорректный ввод числа. Попробуйте ещё раз");
-            
-        }
+        };
+            Console.WriteLine("Введите название вашего товара:");
+            string? product = Console.ReadLine();
 
+            Console.WriteLine("Введите адрес доставки:");
+            string? address = Console.ReadLine();
+
+            Console.WriteLine("\nПодтверждение введенных данных:");
+            Console.WriteLine($"Здравствуйте, {name}, вы заказали {productQuantity} {product} на адрес {address}, верно?");
+            Console.WriteLine("-- да, нет");
+            string answer = Console.ReadLine();
+            switch (answer)
+            {
+                case "да":
+                    Console.WriteLine($"{name}, ваш заказ: {product} в количестве: {productQuantity} оформлен, ожидайте доставку по адресу: {address} к {todayDate}");
+                    break;
+                default:
+                    Console.WriteLine("Пожалуйста, перепроверьте свои данные");
+                    break;
+            }     
     }
 }
