+++ lecture1(14.03)/CommitOrder2/ConfirmOrder.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34622.214
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommitOrder2", "CommitOrder2\CommitOrder2.csproj", "{D83D273D-E9EA-45DF-A06C-95E5A35A8C99}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConfirmOrder", "CommitOrder2\ConfirmOrder.csproj", "{D83D273D-E9EA-45DF-A06C-95E5A35A8C99}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
